api/requirements.txt
fastapi==0.95.1
requests==2.28.1
shroomdk==1.0.2
uvicorn==0.22.0
python-dotenv==1.0.0